% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccr_smreg.R
\name{ccr_smreg}
\alias{ccr_smreg}
\title{Semiparametric Marginal Regression for Clustered Competing Risks Data with Missing Cause of Failure}
\usage{
ccr_smreg(
  data,
  formula11 = y ~ x + Z1 + Z2,
  formula12 = y ~ x + Z1 + Z2,
  formula21 = Surv(x, d) ~ Z1 + Z2,
  formula22 = Surv(x, d) ~ Z1 + Z2,
  w = TRUE,
  var.method = c("None", "CF", "BS"),
  nboot = 1,
  ...
)
}
\arguments{
\item{data}{a data frame in suitable format.}

\item{formula11}{a formula relating the response for cause 1 \code{y} to a set of covariates.}

\item{formula12}{a formula relating the response for cause 1 \code{y} to a set of covariates.}

\item{formula21}{a formula relating the survival object \code{Surv(x, d)} to a set of covariates.}

\item{formula22}{a formula relating the survival object \code{Surv(x, d)} to a set of covariates.}

\item{w}{logical value: if TRUE, the estimation procedure is weighed by cluster size.}

\item{var.method}{a character string specifying the method for variance calcultaion. If None, the variance will
not be calculated; If CF, the close form variance will be used; If BS, the bootstrapping methods will be
used.}

\item{nboot}{a number of bootstrap samples for estimating variances when bootstrapping methods are used}

\item{...}{for future methods}
}
\value{
An object of class \code{ccr_smreg} with components:
\item{est1}{a list for the estimates for cause 1}
\item{est2}{a list for the estimates for cause 2}
\item{se1}{a list for the standard errors for cause 1}
\item{se2}{a list for the standard errors for cause 2}
\item{formula}{a list for the formulas}
\item{x}{a vector for time points where cumulative incidence functions are evaluated}
\item{t}{a vector for time points where cumulative residual processes are evaluated}
\item{Q1}{lower limit for cumulative incidence functions}
\item{Q9}{upper limit for cumulative incidence functions}
\item{q1}{lower limit for cumulative residual process}
\item{q9}{upper limit for cumulative residual process}
\item{nc}{number of clusters}
}
\description{
This function \code{ccr_smreg} conducts semiparametric marginal regression for clustered competing risks data with missing cause of failure. It implements the proposed maximum partial pseudolikelihood estimation method by Zhou, W. et al.(2021) based on semiparametric marginal proportional cause-specific hazards model. The standard errors for the estimated regression coefficients are calculated based on the close-form variance or bootstrapping.
}
\examples{
\dontrun{
library(survival)
library(ClusteredMPPLE)
data("simulated_data")
fit <- ccr_smreg(data=simulated_data, formula11 =  y ~ x + Z1 + Z2, formula12 =  y ~ x + Z1 + Z2,
formula21 = Surv(x, d) ~ Z1 + Z2, formula22 = Surv(x, d) ~ Z1 + Z2,
w = TRUE, var.method = "BS", nboot = 10)
summary(fit)
}
}
\references{
{Zhou, W., Bakoyannis, G., Zhang, Y., & Yiannoutsos, C. T. (2021). Semiparametric Marginal Regression for
Clustered Competing Risks Data with Missing Cause of Failure. arXiv preprint arXiv:2104.09090.}
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
