% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif.R
\name{cif.ccr_smreg}
\alias{cif.ccr_smreg}
\title{Obtain Cumulative Incidence Function and Confidence Band}
\usage{
cif.ccr_smreg(object, band = TRUE, sims = 1, ...)
}
\arguments{
\item{object}{an object of class \code{ccr_smreg}, which is a result of a call to \code{ccr_smreg}.}

\item{band}{logical value: if TRUE, it will calculate and return the 95\% simutanous confidence bands.}

\item{sims}{a number of simulations used for calculating 95\% simutanous confidence bands.}

\item{...}{for future methods.}
}
\value{
An object of class \code{cif.ccr_smreg} with components:
\item{x}{time points}
\item{CIF1}{the cumulative incidence function for cause 1}
\item{CIF2}{the cumulative incidence function for cause 2}
\item{CIF1_band}{a list of 95\% simutanous confidence bands for cause 1}
\item{CIF2_band}{a list of 95\% simutanous confidence bands for cause 2}
}
\description{
\code{cif} method for class \code{ccr_smreg}. It provides the cumulative incidence function with given covariates and provide equal precision bands and Hall-Wellner bands.
}
\examples{
\dontrun{
library(survival)
data("simulated_data")
fit <- ccr_smreg(data=simulated_data, formula11 =  y ~ x + Z1 + Z2, formula12 =  y ~ x + Z1 + Z2,
formula21 = Surv(x, d) ~ Z1 + Z2, formula22 = Surv(x, d) ~ Z1 + Z2,
w = TRUE, var.method = "BS", nboot = 100)
## Continuing the ccr_smreg(...) example
pfit <- cif.ccr_smreg(object = fit, band = TRUE, sims = 1000)
plot(pfit, xlim=c(0.2,1))
}
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
