% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_cif_band_bs.R
\name{clustered_cif_band_bs}
\alias{clustered_cif_band_bs}
\title{Obtain confidence band for cumulative incidence functions based on bootstrap variance}
\usage{
clustered_cif_band_bs(CIF, CIF_boot, range_lower, range_upper, nc, ...)
}
\arguments{
\item{CIF}{a list of time points and cumulative incidence functions.}

\item{CIF_boot}{a matrix of cumulative incidence function bootstrap samples.}

\item{range_lower}{lower limit for cumulative incidence functions.}

\item{range_upper}{upper limit for cumulative incidence functions.}

\item{nc}{number of clusters.}

\item{...}{for future methods.}
}
\value{
A list with components:
\item{LB_EP}{a vector for lower limit of equal precision bands}
\item{UB_EP}{a vector for upper limit of equal precision bands}
\item{LB_HW}{a vector for lower limit of Hall-Wellner bands}
\item{UB_HW}{a vector for upper limit of Hall-Wellner bands}
\item{range}{a vector for range of confidence bands}
}
\description{
This function provides equal precision bands and Hall-Wellner bands for cumulative incidence functions.
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
