% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_cif_band_cf.R
\name{clustered_cif_band_cf}
\alias{clustered_cif_band_cf}
\title{Obtain confidence band for cumulative incidence functions based on close form variance}
\usage{
clustered_cif_band_cf(
  se1,
  se2,
  x,
  range_lower,
  range_upper,
  cause = 1,
  sims = 1000,
  ...
)
}
\arguments{
\item{se1}{a list for the standard errors for cause 1.}

\item{se2}{a list for the standard errors for cause 2.}

\item{x}{time points.}

\item{range_lower}{lower limit for cumulative incidence functions.}

\item{range_upper}{upper limit for cumulative incidence functions.}

\item{cause}{1 or 2, the cause of failure to be evaluated.}

\item{sims}{a number of simulations used for calculating 95\% simutanous confidence bands.}

\item{...}{for future methods.}
}
\value{
A list with components:
\item{V_CIF}{a vector of pointwise confidence intervals}
\item{LB_EP}{a vector for lower limit of equal precision bands}
\item{UB_EP}{a vector for upper limit of equal precision bands}
\item{LB_HW}{a vector for lower limit of Hall-Wellner bands}
\item{UB_HW}{a vector for upper limit of Hall-Wellner bands}
\item{range}{a vector for range of confidence bands}
}
\description{
This function provides equal precision bands and Hall-Wellner bands for cumulative incidence functions.
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
