% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_cif_est.R
\name{clustered_cif_est}
\alias{clustered_cif_est}
\title{Obtain Cumulative Incidence Function Estimates}
\usage{
clustered_cif_est(est1, est2, cause, t, ...)
}
\arguments{
\item{est1}{a list for the estimates for cause 1.}

\item{est2}{a list for the estimates for cause 2.}

\item{cause}{1 or 2, the cause of failure to be evaluated.}

\item{t}{time points.}

\item{...}{for furthur arguments.}
}
\value{
Alist with components:
\item{x}{a vector for time points where cumulative incidence functions are evaluated}
\item{CIF}{a vector for the cumulative incidence function evaluated at time points x}
}
\description{
It provides the cumulative incidence function with given covariates for specific cause.
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
