% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_mpple_est.R
\name{clustered_mpple_est}
\alias{clustered_mpple_est}
\title{Obtain Coefficient Estimates with proposed method using proposed method}
\usage{
clustered_mpple_est(
  data,
  formula1 = y ~ x + Z1 + Z2,
  formula2 = Surv(x, d) ~ Z1 + Z2,
  cause = 1,
  w = TRUE,
  t,
  ...
)
}
\arguments{
\item{data}{a data frame in suitable format.}

\item{formula1}{a formula relating the response for \code{y} to a set of covariates.}

\item{formula2}{a formula relating the survival object \code{Surv(x, d)} to a set of covariates.}

\item{cause}{1 or 2, the cause of failure to be evaluated.}

\item{w}{logical value: if TRUE, the estimation procedure is weighed by cluster size.}

\item{t}{a vector of time points.}

\item{...}{for future methods.}
}
\value{
A list with components:
\item{beta}{a vector of regression coefficients}
\item{time}{a vector of time}
\item{H}{a vector of cumulative hazard function}
\item{crp}{a vector of cumulative residual process}
}
\description{
Obtain coefficient estimates for semiparametric marginal regression for clustered competing risks data with missing cause of failure.
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
