% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_mpple_se_cf.R
\name{clustered_mpple_se_cf}
\alias{clustered_mpple_se_cf}
\title{Obtain coefficient estimates standard error using close form variance}
\usage{
clustered_mpple_se_cf(
  data,
  formula1 = y ~ x + Z1 + Z2,
  formula2 = Surv(x, d) ~ Z1 + Z2,
  cause = 1,
  w = TRUE,
  ...
)
}
\arguments{
\item{data}{a data frame in suitable format.}

\item{formula1}{a formula relating the response for \code{y} to a set of covariates.}

\item{formula2}{a formula relating the survival object \code{Surv(x, d)} to a set of covariates.}

\item{cause}{1 or 2, the cause of failure to be evaluated.}

\item{w}{logical value: if TRUE, the estimation procedure is weighed by cluster size.}

\item{...}{for future methods.}
}
\value{
A list with components:
\item{beta}{a vector of regression coefficients}
\item{SE_B}{a vector of standard errors of regression coefficients}
\item{time}{a vector of time}
\item{H}{a vector of cumulative hazard function}
\item{phi}{useful quantities for incluence function}
\item{dphi}{useful quantities for incluence function}
}
\description{
Obtain coefficient estimates standard error for semiparametric regression for competing risks data
with missing cause of failure using close form variance.
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
