% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_res_band.R
\name{clustered_res_band}
\alias{clustered_res_band}
\title{Obtain goodness-of-fit for cumulative residual process}
\usage{
clustered_res_band(crp, crp_boot, t, range_lower, range_upper, nc, ...)
}
\arguments{
\item{crp}{a vector of cumulative residual process.}

\item{crp_boot}{a matrix of cumulative residual process bootstrap samples.}

\item{t}{a vector of time points.}

\item{range_lower}{lower limit for cumulative residual process.}

\item{range_upper}{upper limit for cumulative residual process.}

\item{nc}{number of clusters.}

\item{...}{for future arguments.}
}
\value{
A list with components:
\item{ca}{a number for limit of goodness-of-fit bands}
\item{p.value}{a number for p-value of goodness-of-fit test}
\item{range_r}{a vector for range of goodness-of-fit bands}
}
\description{
This function provides goodness-of-fit bands for cumulative residual process.
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
