% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp.R
\name{crp.ccr_smreg}
\alias{crp.ccr_smreg}
\title{Obtain Cumulative Residual Process and Goodness-of-fit Confidence Band}
\usage{
crp.ccr_smreg(object, ...)
}
\arguments{
\item{object}{an object of class \code{ccr_smreg}, which is a result of a call to \code{ccr_smreg}.}

\item{...}{for future methods.}
}
\value{
An object of class \code{crp.ccr_smreg} with components:
\item{t}{time points}
\item{res1}{the cumulative residual process for cause 1}
\item{res2}{the cumulative residual process for cause 1}
\item{res1_band}{a list of goodness-of-fit bands and p-value for cause 1}
\item{res2_band}{a list of goodness-of-fit bands and p-value for cause 2}
}
\description{
\code{crp} method for class \code{ccr_smreg}. It provides the cumulative residual process and goodness-of-fit bands.
}
\examples{
\dontrun{
library(survival)
data("simulated_data")
fit <- ccr_smreg(data=simulated_data, formula11 =  y ~ x + Z1 + Z2, formula12 =  y ~ x + Z1 + Z2,
formula21 = Surv(x, d) ~ Z1 + Z2, formula22 = Surv(x, d) ~ Z1 + Z2,
w = TRUE, var.method = "BS", nboot = 100)
## Continuing the ccr_smreg(...) example
rfit <- crp.ccr_smreg(object = fit)
plot(rfit)
}
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
