% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif.R
\name{plot.cif.ccr_smreg}
\alias{plot.cif.ccr_smreg}
\title{Plot a \code{cif.ccr_smreg} object}
\usage{
\method{plot}{cif.ccr_smreg}(x, ...)
}
\arguments{
\item{x}{the result of a call to 'cif.ccr_smreg'}

\item{...}{for future methods}
}
\description{
Plot the cumulative incidence functions and 95\% simutanous confidence bands from a \code{cif.ccr_smreg} object.
}
\examples{
\dontrun{
plot(pfit)
}
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
