% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp.R
\name{plot.crp.ccr_smreg}
\alias{plot.crp.ccr_smreg}
\title{Plot a \code{crp.ccr_smreg} object}
\usage{
\method{plot}{crp.ccr_smreg}(x, cause = 1, ...)
}
\arguments{
\item{x}{the result of a call to 'crp.ccr_smreg'}

\item{cause}{1 or 2, the cause of failure to be evaluated}

\item{...}{for future methods}
}
\description{
Plot the cumulative residual process from a \code{crp.ccr_smreg} object.
}
\examples{
\dontrun{
plot(rfit)
}
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
