% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccr_smreg.R
\name{print.summary.ccr_smreg}
\alias{print.summary.ccr_smreg}
\title{Print method for \code{summary.ccr_smreg} objects}
\usage{
\method{print}{summary.ccr_smreg}(x, ...)
}
\arguments{
\item{x}{the result of a call to 'summary.ccr_smreg'}

\item{...}{for future methods}
}
\description{
Produce a printed summary of the fitted semiparametric marginal regression model
}
\author{
Wenxian Zhou, \email{wz11 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
