% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_HIV_data.R
\docType{data}
\name{pseudo_HIV_data}
\alias{pseudo_HIV_data}
\title{pseudo_HIV_data data set}
\format{
A data.frame with 7 columns:
\describe{
  \item{clusterid}{Cluster ID, number from 1 to the number of clusters}
  \item{x}{Observed time}
  \item{c}{Cause of failure, 1 = cause 1, 2 = cause 2, 0 = right censoring, 99 = missing cause of failure}
  \item{Sex}{Covariate Sex}
  \item{Age}{Covariate Age}
  \item{CD4}{Covariate CD4}
  \item{HIV}{Covariate HIV}
}
}
\source{
Simulated Data
}
\usage{
pseudo_HIV_data
}
\description{
We generate sample HIV dataset for this package
}
\examples{
library(ClusteredMPPLE)
data(pseudo_HIV_data)
}
\keyword{datasets}
