% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{simulated_data data set}
\format{
A data.frame with 7 columns
}
\source{
Simulated Data
}
\usage{
simulated_data
}
\description{
We simulated a sample dataset with 50 cluster size for this package
}
\examples{
library(ClusteredMPPLE)
data(simulated_data)
}
\keyword{datasets}
