% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_HIV_data.R
\docType{data}
\name{pseudo_HIV_data}
\alias{pseudo_HIV_data}
\title{pseudo_HIV_data data set}
\format{
A data.frame with 9 columns
}
\source{
Simulated Data
}
\usage{
pseudo_HIV_data
}
\description{
We generate sample HIV dataset for this package
}
\examples{
library(ClusteredMPPLE)
data(pseudo_HIV_data)
}
\keyword{datasets}
