% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{simulated_data data set}
\format{
A data.frame with 6 columns:
\describe{
  \item{clusterid}{Cluster ID, 1 to number of clusters}
  \item{x}{Observed time}
  \item{c}{Cause of failure, 1 = cause 1, 2 = cause 2, 0 = right censoring}
  \item{r}{Missing indicator, 1 if observed, 0 if missing}
  \item{Z1}{First covariate}
  \item{Z2}{Second covariate}
}
}
\source{
Simulated Data
}
\usage{
simulated_data
}
\description{
We simulated a sample dataset with 50 cluster size for this package
}
\examples{
library(ClusteredMPPLE)
data(simulated_data)
}
\keyword{datasets}
